Advances for #81937: We always sanitize names when creating geoserver utils, to ensure that no badly named elements are ever created in geoserver.
Note that this means that if you don't apply the sanitizing function for stored references to geoserver object names in your database, geoserver will not find the created elements as they won't be named the same.

tl;dr: Sanitize your names in client code even if we are sanitizing in the dao for better results.